requirements.txt
aiohttp==3.9.3
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pyyaml==6.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
black==24.1.1
ruff==0.2.1
mypy==1.8.0
types-python-dateutil==2.8.19.20240106
types-PyYAML==6.0.12.12
aiofiles==23.2.1
structlog==24.1.0
typing-extensions==4.9.0
coverage==7.4.1
aioresponses==0.7.6
freezegun==1.4.0 